[Modify, Fix] 뒤로 가기를 해도 모달 창이 열려있는 현상 수정 등 일부 버그 수정 (2023/09/19)
