+++ src/App.jsx
@@ -1,32 +1,30 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { AnimatePresence } from 'framer-motion';
 import { HelmetProvider } from 'react-helmet-async';
 import { Toaster } from 'react-hot-toast';
 import { RouterProvider } from 'react-router-dom';
-import { AuthProvider } from './contexts/AuthProvider.jsx';
-import ChannelProvider from './contexts/Channel.jsx';
-import ReplyProvider from './contexts/Reply.jsx';
+import { AuthProvider, ChannelProvider, ReplyProvider } from './contexts';
 import router from './routes.jsx';
-import { AnimatePresence } from 'framer-motion';
 
 const queryClient = new QueryClient();
 
 function App() {
   return (
     <>
-      <AnimatePresence>
-        <HelmetProvider>
-          <QueryClientProvider client={queryClient}>
-            <AuthProvider>
-              <ChannelProvider>
-                <ReplyProvider>
+      <HelmetProvider>
+        <QueryClientProvider client={queryClient}>
+          <AuthProvider>
+            <ChannelProvider>
+              <ReplyProvider>
+                <AnimatePresence>
                   <RouterProvider router={router} />
-                </ReplyProvider>
-              </ChannelProvider>
-            </AuthProvider>
-          </QueryClientProvider>
-        </HelmetProvider>
-        <Toaster />
-      </AnimatePresence>
+                </AnimatePresence>
+              </ReplyProvider>
+            </ChannelProvider>
+          </AuthProvider>
+        </QueryClientProvider>
+      </HelmetProvider>
+      <Toaster />
     </>
   );
 }
