+++ src/components/button/WritingButton.jsx
@@ -10,7 +10,7 @@ function WritingButton() {
       onClick={() => {
         navigate('/writing');
       }}
-      className="flex h-[34px] w-[88px] items-center justify-center gap-2 rounded-lg bg-lionly-secondary-color text-lionly-md text-lionly-white shadow-lg"
+      className="flex h-[34px] w-[88px] items-center justify-center gap-2 rounded-lg bg-lionly-secondary-color text-lionly-md text-lionly-white shadow-lg transition-all hover:scale-125 focus:scale-125"
     >
       <PenSVG aria-hidden />
       글쓰기
