+++ src/components/content/index.js
@@ -2,3 +2,4 @@ export { default as Header } from './Header.jsx';
 export { default as Content } from './Content.jsx';
 export { default as Comments } from './Comments.jsx';
 export { default as InsertComment } from './InsertComment.jsx';
+export { default as ModalReply } from './ModalReply.jsx';
