+++ src/components/feed/index.js
@@ -1,0 +1,3 @@
+export { default as ChannelTab } from './ChannelTab.jsx';
+export { default as FeedHeader } from './FeedHeader.jsx';
+export { default as FeedList } from './FeedList.jsx';
