+++ src/utils/index.js
@@ -1,5 +1,5 @@
 export { default as getNodeIndex } from './getNodeIndex';
 export { default as getPbImageURL } from './getPbImageURL';
 export { default as calcTimeDifference } from './calcTimeDifference';
-export { default as handleTabArrowControl } from './handleTabArrowControl';
+export { default as handleKeyboardArrowControl } from './handleKeyboardArrowControl';
 export { default as getDate } from './getDate';
